Split the overlays back out of canvas.tsx
The earlier split lost its footing when main landed: the merge took main's
canvas.tsx whole, so crop-overlay.tsx and selection-overlay.tsx sat there
importing into nobody while canvas.tsx carried all three overlays again.

Redone from main's text as pure moves, with the two new arrivals placed the
way the old split placed the rest: SmartGuides joins selection-overlay,
since it draws the same screen-space chrome over the same gesture and reads
nothing the file doesn't already have.

scripts/test-snap.ts arrived with its own copy of the tally; it uses the
shared harness now, like every other suite.
